Fix 34 bugs found by full QA audit across all 14 crates
Critical fixes:
- ferray-core: validate non-negative strides in broadcast_to to prevent UB
- ferray-linalg: hoist data collection outside from_fn closure in array2_to_faer (was O(n^4))
- ferray-io: handle Fortran-order for complex32/complex64 dynamic loading
- ferray umbrella: remove non-additive no_std feature that broke --all-features

High-priority fixes:
- ferray-core: iinfo::<u128>().max now uses u128 (was wrapping to -1), view_cast checks alignment
- ferray-ufunc: nextafter/spacing use IEEE 754 bit manipulation, logaddexp/logaddexp2 propagate NaN,
  gcd/lcm/unwrap guard against NaN infinite loops, frexp uses modular reduction
- ferray-linalg: slogdet uses LU decomposition (avoids overflow), inv/solve detect singular matrices
  via NaN/Inf check, cond accepts non-square matrices for SVD-based norms
- ferray-stats: histogram density parameter implemented, sort(axis=None) returns flat IxDyn,
  min/max propagate NaN consistently, SIMD dispatch uses transmute_copy
- ferray-io: checked_mul for shape overflow (DoS), 1MB header_len cap, TOCTOU fix in memmap
- ferray-ma: masked_binary_op validates shapes, min/max propagate NaN
- ferray-polynomial: numerically stable quadratic formula, degree() uses exact zero (not epsilon)
- ferray-random: Laplace rejects u.abs()>=0.5, Poisson squeeze returns k, binomial uses powf not powi
- ferray-strings: search/boolean ops preserve input dimension (was flattening to Ix1)
- ferray-fft: remove unnecessary unsafe impl Send/Sync for FftPlan
- ferray-stride-tricks: reject negative strides in sliding_window_view
- ferray-window: kaiser accepts negative beta (I0 is even)
- ferray umbrella: forward f16/bf16 features, re-export indexing module, remove panic in config

+++ ferray-core/src/dtype/finfo.rs
@@ -74,8 +74,8 @@ impl fmt::Display for FloatInfo {
 pub struct IntInfo {
     /// The smallest representable value.
     pub min: i128,
-    /// The largest representable value.
-    pub max: i128,
+    /// The largest representable value (u128 to accommodate u128::MAX).
+    pub max: u128,
     /// Number of bits in the type.
     pub bits: u32,
 }
@@ -180,7 +180,7 @@ macro_rules! impl_int_type {
             fn int_info() -> IntInfo {
                 IntInfo {
                     min: <$ty>::MIN as i128,
-                    max: <$ty>::MAX as i128,
+                    max: <$ty>::MAX as u128,
                     bits: $bits,
                 }
             }
@@ -239,7 +239,7 @@ pub fn finfo<T: FloatType>() -> FloatInfo {
 /// use ferray_core::dtype::finfo::iinfo;
 /// let info = iinfo::<i32>();
 /// assert_eq!(info.min, i32::MIN as i128);
-/// assert_eq!(info.max, i32::MAX as i128);
+/// assert_eq!(info.max, i32::MAX as u128);
 /// ```
 pub fn iinfo<T: IntType>() -> IntInfo {
     T::int_info()
@@ -282,7 +282,7 @@ mod tests {
     fn iinfo_i32() {
         let info = iinfo::<i32>();
         assert_eq!(info.min, i32::MIN as i128);
-        assert_eq!(info.max, i32::MAX as i128);
+        assert_eq!(info.max, i32::MAX as u128);
         assert_eq!(info.bits, 32);
     }
 
@@ -298,7 +298,7 @@ mod tests {
     fn iinfo_i64() {
         let info = iinfo::<i64>();
         assert_eq!(info.min, i64::MIN as i128);
-        assert_eq!(info.max, i64::MAX as i128);
+        assert_eq!(info.max, i64::MAX as u128);
         assert_eq!(info.bits, 64);
     }
 
@@ -314,7 +314,7 @@ mod tests {
     fn iinfo_u128() {
         let info = iinfo::<u128>();
         assert_eq!(info.min, 0);
-        assert_eq!(info.max, u128::MAX as i128);
+        assert_eq!(info.max, u128::MAX);
         assert_eq!(info.bits, 128);
     }
